feat: 쿠렌토 n:m 테스트 페이지 생성
