+++ src/hooks/use-transform-user.ts
@@ -1,7 +1,7 @@
 import { TGroupFetchMemberInfo } from '@api/group/group-request.type';
 import { TChatSocketType } from '@pages/meeting-room/types';
 import { useGroupStore } from '@stores/group';
-import { useMeetingStore } from '@stores/meeting';
+// import { useMeetingStore } from '@stores/meeting';
 import { useGroupMemberQuery } from './react-query/use-query-group';
 
 const useTransformUser = (socketMsg: TChatSocketType) => {
