+++ src/pages/kurento-service/kurento-service.style.css
@@ -88,8 +88,8 @@
 
 #leaveBtn {
   position: absolute;
-  width: 510px;
-  height: 50px;
+  width: 110px;
+  height: 20px;
   display: flex;
   font-size: 20px;
   top: 90%;
@@ -101,11 +101,12 @@
   display: flex;
   align-items: center;
   justify-content: center;
-  background-color: #013d04;
+  background-color: rgba(0, 0, 0, 0.8);
   color: white;
   font-weight: bold;
   cursor: pointer;
-  visibility: hidden;
+  /* margin-top: 300px; */
+  /* visibility: hidden; */
 }
 #leaveBtn:hover {
   background-color: #002602;
