+++ src/pages/kurento-service/utils/kurento-service.ts
@@ -1,38 +1,89 @@
-/* eslint-disable no-console */
-import { Client } from '@stomp/stompjs';
-import SockJS from 'sockjs-client';
+// eslint-disable-next-line @typescript-eslint/ban-ts-comment
+// @ts-nocheck
 
-const WebSocketService = (() => {
-  let client: Client | null = null;
+const PARTICIPANT_MAIN_CLASS = 'participant main';
+const PARTICIPANT_CLASS = 'participant';
 
-  const connect = () => {
-    client = new Client({
-      webSocketFactory: () => {
-        return new SockJS('https://api.effi.club/signal/connect');
-      },
-      onConnect: () => {
-        console.log('WebSocket 연결됨');
-      },
-      onStompError: (frame) => {
-        console.error('Broker reported error: ' + frame.headers['message']);
-        console.error('Additional details: ' + frame.body);
-      },
-    });
+class Participant {
+  constructor(userId, sendMessage) {
+    this.userId = userId;
 
-    client.activate();
-  };
+    this.container = document.createElement('div');
+    this.container.className = this.isPresentMainParticipant() ? PARTICIPANT_CLASS : PARTICIPANT_MAIN_CLASS;
+    this.container.id = userId;
+    this.span = document.createElement('span');
+    this.video = document.createElement('video');
+    this.rtcPeer = null;
+    this.sendMessage = sendMessage;
+    this.onIceCandidate = this.onIceCandidate.bind(this);
 
-  const disconnect = () => {
-    if (client !== null) {
-      console.log('WebSocket 연결 끊어짐');
-      client.deactivate();
+    this.container.appendChild(this.video);
+    this.container.appendChild(this.span);
+    this.container.onclick = this.switchContainerClass.bind(this);
+    document.getElementById('participants').appendChild(this.container);
+
+    this.span.appendChild(document.createTextNode(userId));
+
+    this.video.id = 'video-' + userId;
+    this.video.autoplay = true;
+    this.video.controls = false;
+  }
+
+  getElement() {
+    return this.container;
+  }
+
+  getVideoElement() {
+    return this.video;
+  }
+
+  switchContainerClass() {
+    if (this.container.className === PARTICIPANT_CLASS) {
+      const elements = Array.prototype.slice.call(document.getElementsByClassName(PARTICIPANT_MAIN_CLASS));
+      elements.forEach(function (item) {
+        item.className = PARTICIPANT_CLASS;
+      });
+
+      this.container.className = PARTICIPANT_MAIN_CLASS;
+    } else {
+      this.container.className = PARTICIPANT_CLASS;
     }
-  };
+  }
 
-  return {
-    connect,
-    disconnect,
-  };
-})();
+  isPresentMainParticipant() {
+    return document.getElementsByClassName(PARTICIPANT_MAIN_CLASS).length !== 0;
+  }
 
-export default WebSocketService;
+  offerToReceiveVideo(error, offerSdp, wp) {
+    if (error) return console.error('sdp offer error');
+    console.log('Invoking SDP offer callback function');
+
+    const msg = {
+      id: 'receiveVideoFrom',
+      sender: this.userId,
+      sdpOffer: offerSdp,
+    };
+    this.sendMessage(msg);
+  }
+
+  onIceCandidate(candidate, wp) {
+    console.log('Local candidate' + JSON.stringify(candidate));
+
+    const message = {
+      id: 'onIceCandidate',
+      userId: this.userId,
+      candidate: candidate,
+    };
+    this.sendMessage(message);
+  }
+
+  dispose() {
+    console.log('Disposing participant ' + this.userId);
+    if (this.rtcPeer) {
+      this.rtcPeer.dispose();
+    }
+    this.container.parentNode.removeChild(this.container);
+  }
+}
+
+export default Participant;
