+++ src/pages/meeting-room/index.tsx
@@ -73,6 +73,7 @@ const MeetingRoom = () => {
             isMeetingFinished={isMeetingFinished}
           />
         </S.Nav>
+
         <S.RoomCameraContainer>
           <S.RoomCameraBox>
             {participatedMember.map((member, idx) => (
@@ -80,11 +81,13 @@ const MeetingRoom = () => {
             ))}
           </S.RoomCameraBox>
         </S.RoomCameraContainer>
+
         <S.RoomButtonContainer className="room-button-container">
           {ROOM_BUTTONS.map((btn, idx) => (
             <RoomButton key={idx} type={btn.type} initialImg={btn.initialImg} changedImg={btn.changedImg} />
           ))}
         </S.RoomButtonContainer>
+        
         <ForceQuitToast isToastOpen={isToastOpen} isToastAnimClose={isToastAnimClose} />
       </S.LeftSection>
       <S.RightSection>
